2nd Place/requirements.txt
rasterio==1.0.24
pandas==0.23.1
pyproj==1.9.6
geopandas==0.6.0
opencv-python>=4.1.2.30
Shapely==1.6.4.post1
scikit-learn==0.21.2
scipy==1.1.0
pyvips==2.0.4
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
albumentations>=0.4.3
efficientnet==1.0.0
image-classifiers==1.0.0b1
kito==1.0.3
catboost==0.12.2
xgboost==0.90
lightgbm==2.3.1